% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{arm_fit}
\alias{arm_fit}
\title{Fitter Function for the Augmented Regression Method}
\usage{
arm_fit(y, x, iter = 10, ...)
}
\arguments{
\item{y}{vector of observations of length \code{n}, or a matrix with
    \code{n} rows.}

\item{x}{design matrix of dimension \code{n * p}.}

\item{iter}{maximum number of bias-correction iterations (\code{K = 10} in the
paper); iteration stops earlier if the corrected VAR becomes non-stationary.}

\item{...}{currently unused.}
}
\description{
Basic function called by \code{arm}. Should only be used directly by experienced
users.
}
\examples{
arm_fit(kms$Ret, as.matrix(kms$DP))$tstat
}
