% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c_low_crit}
\alias{c_low_crit}
\title{Critical values for the lower bound c}
\format{An object of class \code{matrix} with 61 rows and 40 columns.}
\usage{
c_low_crit
}
\description{
Critical values for the lower bound c
}
\keyword{datasets}
