% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c_up_crit}
\alias{c_up_crit}
\title{Critical values for the upper bound c}
\format{An object of class \code{matrix} with 61 rows and 40 columns.}
\usage{
c_up_crit
}
\description{
Critical values for the upper bound c
}
\keyword{datasets}
