% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datam}
\alias{datam}
\title{Monthly data}
\format{A data.frame with 1680 observations
\describe{
  \item{sp.500.ratio}{S and P 500 Price-Dividen Ratio}
 }}
\source{
\url{http://www.econ.yale.edu/~shiller/data.htm}
}
\usage{
datam
}
\description{
year and month
 dividend payout ratio (DE)
long-term yield (LTY)
dividend yield (DY)
dividend-price ratio (DP)
T-bill rate (TBL)
earnings-price ratio (EP)
book-to-market value ratio (BM)
inflation rate (INF)
default yield spread (DFY)
net equity expansion (NTIS)
term spread (TMS)
S&P 500 value-weighted log excess returns (LOG EXCESS VW)
}
\keyword{datasets}
