% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx.R
\name{delta}
\alias{delta}
\title{Finding the delta coefficient}
\usage{
delta(object, mat = F, diag = FALSE)
}
\arguments{
\item{object}{class "ivx"}

\item{mat}{matrix format}

\item{diag}{include diagonal elements}
}
\description{
Finding the delta coefficient
}
