% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-texreg-methods.R
\name{extract.ivx}
\alias{extract.ivx}
\alias{extract.ivx_ar}
\title{\code{\link{extract}} method for \code{ivx} objects}
\usage{
extract.ivx(
  model,
  include.wald = TRUE,
  inlude.nobs = TRUE,
  include.aic = FALSE,
  include.bic = FALSE,
  include.rsquared = FALSE,
  include.adjrs = FALSE
)

extract.ivx_ar(
  model,
  include.wald = TRUE,
  inlude.nobs = TRUE,
  include.aic = FALSE,
  include.bic = FALSE,
  include.rsquared = FALSE,
  include.adjrs = FALSE
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.wald}{Report the Wald statistic.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.rsquared}{Report the R-squared.}

\item{include.adjrs}{Report the Adjusted R-squared.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{ivx} objects
}
