% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx-package.R
\docType{package}
\name{ivx-package}
\alias{ivx-package}
\title{Robust Econometric Inference}
\description{
Drawing statistical inference on the coefficients of a short- or long-horizon
predictive regression with persistent regressors by using the IVX method of
Magdalinos and Phillips (2009) DOI: 10.1017/S0266466608090154 and
Kostakis, Magdalinos and Stamatogiannis (2015) DOI: 10.1093/rfs/hhu139.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvasilopoulos/ivx}
  \item Report bugs at \url{https://github.com/kvasilopoulos/ivx/issues}
}

}
\author{
\strong{Maintainer}: Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}

Authors:
\itemize{
  \item Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}
  \item Efthymios Pavlidis \email{e.pavlidis@lancaster.ac.uk}
}

}
\keyword{internal}
