% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx_roll.R
\name{ivx_roll_fit}
\alias{ivx_roll_fit}
\title{Fitter Functions for Rolling IVX Models}
\usage{
ivx_roll_fit(y, x, horizon = 1, minw = NULL, offset = NULL, ...)
}
\arguments{
\item{y}{vector of observations of length \code{n}, or a matrix with
    \code{n} rows.}

\item{x}{design matrix of dimension \code{n * p}.}

\item{horizon}{is the horizon (default horizon = 1 corresponds to a
short-horizon regression).}

\item{minw}{minimum window}

\item{offset}{(numeric of length \code{n}).  This can be used to
    specify an \emph{a priori} known component to be included in the
    linear predictor during fitting.}

\item{...}{currently disregarded.}
}
\description{
Basic function called by \code{ivx_roll} to fit predictive models.
These should only be used directly by experienced users.
}
\examples{
ivx_fit(monthly$Ret, as.matrix(monthly$LTY))
}
