% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kms_quarterly}
\alias{kms_quarterly}
\title{KMS Quarterly data}
\format{
A data.frame with 13 variables and 345 observations.
}
\source{
\url{https://drive.google.com/open?id=1FdT2STHO2Lnlweom4AwICVf-rpVMfgV4}
}
\usage{
kms_quarterly
}
\description{
The dataset that was used in Kostakis et al (2015), containing
a range of financial variables.

\itemize{
\item Date: year-month-date
\item DE: dividend payout ratio
\item LTY: long-term yield
\item DY: dividend yield
\item DP: dividend-price ratio
\item TBL: T-bill rate
\item EP: earnings-price ratio
\item BM: book-to-market value ratio
\item INF: inflation rate
\item DFY: default yield spread
\item NTIS: net equity expansion
\item TMS: term spread
\item Ret: S&P 500 value-weighted log excess returns
}
}
\keyword{datasets}
