% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly}
\alias{monthly}
\title{Monthly dataset of KMS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1033 rows and 13 columns.
}
\usage{
monthly
}
\description{
Monthly dataset of KMS
}
\keyword{internal}
