% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonf.R
\name{qtest}
\alias{qtest}
\title{Campel Yogo bonferroni confidence intervals Q-test}
\usage{
qtest(y, x, method = c("bonferroni", "infeasible", "sup"), cest = NULL,
  sig = 0.1)
}
\arguments{
\item{y}{the regressand}

\item{x}{the regressor}

\item{method}{c("bonferroni", "infeasible", "sup")}

\item{cest}{for sup test the user have to provide an estimation of c}

\item{sig}{10\%}
}
\value{
the upper and lower bound of the regression coefficient and the
autoregressive coefficient
}
\description{
Campel Yogo bonferroni confidence intervals Q-test
}
\examples{
x <- cumsum(rnorm(100))
y <- 2*x + rnorm(100)
qtest(y, x)
}
