% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quarterly}
\alias{quarterly}
\title{Quarterly dataset of KMS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 345 rows and 13 columns.
}
\usage{
quarterly
}
\description{
Quarterly dataset of KMS
}
\keyword{internal}
