% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx-ar.R
\name{summary.ivx_ar}
\alias{summary.ivx_ar}
\alias{print.summary.ivx_ar}
\title{Summarizing IVX-AR Model Fits}
\usage{
\method{summary}{ivx_ar}(object, ...)

\method{print}{summary.ivx_ar}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{object of class "ivx_ar", usually, a result of a call to ivx_ar.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"summary.lm"}, usually, a result of a
    call to \code{summary.lm}.}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If \code{TRUE}, \sQuote{significance stars}
    are printed for each coefficient.}
}
\description{
summary method for class "ivx".
}
\examples{
mod <- ivx_ar(Ret ~ LTY, data = kms)

summary(mod)
}
