% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroot.R
\name{uroot}
\alias{uroot}
\title{Unit root and stationarity test overview}
\usage{
uroot(y, model = c("constant", "trend"), lag.max = 4)
}
\arguments{
\item{y}{dataset}

\item{model}{include constant and or trend}

\item{lag.max}{maximum number of lags}
}
\description{
Unit root and stationarity test overview
}
\examples{
names <- datam \%>\%
  select(-Date, -Ret) \%>\%
  names()

ur <- cbind(names, datam \%>\%
           select(-Date, -Ret) \%>\%
           map(uroot) \%>\%
           reduce(rbind))

}
