% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ylpc}
\alias{ylpc}
\title{YLPC Quarterly data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 174 rows and 12 columns.
}
\source{
\url{https://www.tandfonline.com/doi/suppl/10.1080/01621459.2019.1686392/suppl_file/uasa_a_1686392_sm7226.zip}
}
\usage{
ylpc
}
\description{
The dataset that was used in ..., containing
a range of  variables.
}
\keyword{datasets}
