% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_test.R
\name{ac_test}
\alias{ac_test}
\title{If p-value < 0.051: You can reject the null hypothesis assuming a 5% chance of making a mistake.
So you can assume that your values are showing dependence on each other.}
\usage{
ac_test(x, max = 5)
}
\description{
x <- ivx(hpi ~ cpi + def + int + log(res), data = ylpc)
lmtest::bgtest(hpi ~ cpi + def + int + log(res), data = ylpc)
ac_test(x, 5)
ac_test.ivx(x)
t2 <- ivx_ar(hpi ~ cpi + def + int + res, data = ylpc)
}
