% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_test.R
\name{ac_test}
\alias{ac_test}
\title{Autocorrelation tests}
\usage{
ac_test(x, lag_max = 5)
}
\arguments{
\item{x}{the residuals or an \code{ivx} object.}

\item{lag_max}{the maximum length of lags.}
}
\description{
Autocorrelation tests
}
\examples{
obj <- ivx(hpi ~ cpi + def + int + log(res), data = ylpc)
lmtest::bgtest(hpi ~ cpi + def + int + log(res), data = ylpc)
ac_test(obj, 5)

}
