% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_test.R
\name{ac_test_lb}
\alias{ac_test_lb}
\alias{ac_test_bp}
\alias{ac_test_bg.default}
\alias{ac_test}
\title{Autocorrelation tests}
\usage{
ac_test_lb(x, lag)

ac_test_bp(x, lag)

ac_test_bg.default(x, order = 1, type = c("Chisq", "F"), fill = 0)

ac_test(x, lag_max = 5)
}
\arguments{
\item{x}{the residuals or an \code{ivx} object.}

\item{lag}{the number of lags.}

\item{order}{lag TODO}

\item{type}{the type of test statistic to be returned. Either "Chisq" for
the Chi-squared test statistic or "F" for the F test statistic.}

\item{fill}{starting values for the lagged residuals in the auxiliary regression.
By default 0 but can also be set to NA.}

\item{lag_max}{the maximum length of lags.

obj <- ivx(hpi ~ cpi + def + int + log(res), data = ylpc)
lmtest::bgtest(hpi ~ cpi + def + int + log(res), data = ylpc)
ac_test(obj, 5)
ac_test.ivx(x)
t2 <- ivx_ar(hpi ~ cpi + def + int + log(res), data = ylpc)}
}
\description{
Autocorrelation tests
}
