% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_test.R
\name{ac_test_}
\alias{ac_test_}
\alias{ac_test_wald}
\title{Tests  for autocorrelation}
\usage{
ac_test_wald(x, lag)
}
\arguments{
\item{x}{an \code{ivx} model or a \verb{numeric vector}, usually the residuals from an ols regression.}

\item{lag}{the number of lags.}
}
\value{
a numeric scalar or numeric vector.
}
\description{
\itemize{
\item \code{ac_test_wald} Wald test
\item \code{ac_test_lb} Ljung-Box
\item \code{ac_test_bp} Box-Pierce
\item \code{ac_test_bg}Breusch-Godfrey
}
}
\details{
#' If p-value < 0.051: You can reject the null hypothesis assuming a
5\% chance of making a mistake. So you can assume that your values are showing
dependence on each other.
}
\examples{

mdl <- ivx(hpi ~ cpi + inv, data = ylpc)
ac_test_wald(mdl)


}
\seealso{
\code{Box.test} \code{lmtest::bgtest}
}
