% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx.R
\name{delta}
\alias{delta}
\title{Calculate the delta coefficient}
\usage{
delta(object)
}
\arguments{
\item{object}{class "ivx"}
}
\value{
This should have row and column names corresponding to the parameter
names given by the coef method.
}
\description{
Calculate the correlation coefficient between the residuals of regression between
the predictive and the the autoregressive regression.
}
