% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{delta}
\alias{delta}
\title{Calculate the delta coefficient}
\usage{
delta(object)
}
\arguments{
\item{object}{on object of class "ivx"}
}
\value{
A vector of the estimated correlation coefficients. This should have
row and column names corresponding to the parameter names given by the coef method.
}
\description{
Computes the long-run correlation coefficient between the residuals of the
predictive regression and the autoregressive model for the regressor.
}
\examples{
mod <- ivx(Ret ~ LTY, data = monthly)

delta(mod)
}
