% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-texreg-methods.R
\name{extract.ivx}
\alias{extract.ivx}
\title{\code{\link{extract}} method for \code{ivx} objects}
\usage{
extract.ivx(
  model,
  include.wald = TRUE,
  inlude.nobs = TRUE,
  include.rsquared = FALSE,
  include.adjrs = FALSE
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.pvalues}{Report p-values?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{Arima} objects created by the
\code{\link[stats]{arima}} function in the \pkg{stats} package.
}
