% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx-package.R
\docType{package}
\name{ivx-package}
\alias{ivx}
\alias{ivx-package}
\title{Robust Econometric Inference}
\description{
Conducting inference on the regression coefficient with a highly persistent regressor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvasilopoulos/ivx}
  \item Report bugs at \url{https://github.com/kvasilopoulos/ivx/issues}
}

}
\author{
\strong{Maintainer}: Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}

Authors:
\itemize{
  \item Eftymios Pavlidis \email{e.pavlidis@lancaster.ac.uk}
}

}
\keyword{internal}
