% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx.R
\name{ivx}
\alias{ivx}
\alias{ivx_fit}
\alias{print.ivx}
\title{Fitting IVX Models}
\usage{
ivx(formula, data, horizon, subset, na.action, model = TRUE,
  contrasts = NULL, offset, ...)

ivx_fit(y, x, h = 1, offset = NULL, ...)

\method{print}{ivx}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{lm} is called.}

\item{horizon}{is the horizon (default horizon=1 corresponds to a
short-horizon regression)}

\item{subset}{an optional vector specifying a subset of observations
    to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is set by
    the \code{na.action} setting of \code{\link{options}}, and is
    \code{\link{na.fail}} if that is unset.  The \sQuote{factory-fresh}
    default is \code{\link{na.omit}}.  Another possible value is
    \code{NULL}, no action.  Value \code{\link{na.exclude}} can be useful.}

\item{model}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }

\item{contrasts}{an optional list. See the \code{contrasts.arg}
    of \code{\link{model.matrix.default}}.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector of length equal to
    the number of cases.  One or more \code{\link{offset}} terms can be
    included in the formula instead or as well, and if more than one are
    specified their sum is used.  See \code{\link{model.offset}}.}

\item{...}{additional arguments to be passed to the low level
    regression fitting functions (see below).}

\item{y}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }

\item{x}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }
}
\value{
ivx returns an object of class "ivx".

An object of class "ivx" is a list containing the following components:
\item{coefficients_ols}{coefficients}
\item{coefficients_ivx}{coefficients}
\item{call}{call}
\item{terms}{terms}
\item{model}{ model}
}
\description{
Fitting IVX Models
}
\references{
Phillips, P. C., & Magdalinos, T. (2009). Econometric inference
in the vicinity of unity. Singapore Management University, CoFie Working
Paper, 7.

Kostakis, A., Magdalinos, T., & Stamatogiannis, M. P. (2014).
Robust econometric inference for stock return predictability. The Review of
Financial Studies, 28(5), 1506-1553.

Phillips, P. C., & Lee, J. H. (2013). Predictive regression under
various degrees of persistence and robust long-horizon regression. Journal
of Econometrics, 177(2), 250-264.
}
