% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx.R
\name{ivx_fit}
\alias{ivx_fit}
\title{Fitter Functions for ivx Models}
\usage{
ivx_fit(y, x, horizon = 1, offset = NULL, ...)
}
\arguments{
\item{y}{vector of observations of length \code{n}, or a matrix with
    \code{n} rows.}

\item{x}{design matrix of dimension \code{n * p}.}

\item{horizon}{is the horizon (default horizon = 1 corresponds to a
short-horizon regression)}

\item{offset}{(numeric of length \code{n}).  This can be used to
    specify an \emph{a priori} known component to be included in the
    linear predictor during fitting.}

\item{...}{currently disregarded.}
}
\description{
These are the basic computing engines called by `ivx` to fit models.
These should usually not be used directly unless by experienced users.
}
