% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ylpc}
\alias{ylpc}
\title{YLPC Quarterly data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 174 rows and 12 columns.
}
\source{
Yang, B., Long, W., Peng, L., and Cai, Z. (2020).
Autoregressive-model-based inference for predictive regressions with
persistent predictors. Journal of the American Statistical Association,
115(530), 1029-1039. DOI: 10.1080/01621459.2019.1686392
}
\usage{
ylpc
}
\description{
The dataset that was used in ..., containing
a range of  variables.
}
\keyword{datasets}
